+++ src/fullctl/django/models/abstract/service_bridge.py
@@ -160,7 +160,7 @@ class ServiceBridgeReferenceModel(HandleRefModel):
                     pass
 
                 changed = True
-                # print(f"{src_field} changed from {src_value} to {dest_value}")
+                print(f"{src_field} changed from {src_value} to {dest_value}")
                 setattr(self, src_field, dest_value)
 
         # print("sync from reference", self, changed)