+++ src/fullctl/django/models/concrete/__init__.py
@@ -1,4 +1,5 @@
 from fullctl.django.models.concrete.account import *  # noqa: F401, F403
 from fullctl.django.models.concrete.auditlog import *  # noqa: F401, F403
+from fullctl.django.models.concrete.file import *  # noqa: F401, F403
 from fullctl.django.models.concrete.meta import *  # noqa: F401, F403
 from fullctl.django.models.concrete.tasks import *  # noqa: F401, F403