+++ tests/django_tests/testapp/models.py
@@ -2,6 +2,9 @@
 Contains a bunch of concrete models to help testing
 """
 
+from django.db import models
+
+import fullctl.django.models.abstract.meta as meta
 from fullctl.django.models import Task
 from fullctl.django.tasks import qualifiers, register
 
@@ -108,3 +111,77 @@ class FailingTask(Task):
 
     def run(self, *args, **kwargs):
         return int("string")
+
+
+class Data(meta.Data):
+    """
+    Concrete implementation of Data model
+    """
+
+    target = models.CharField(max_length=255)
+
+    class Meta:
+        verbose_name = "Data"
+        verbose_name_plural = "Data"
+        db_table = "test_meta_data"
+
+    class Config:
+        type = "test"
+        source_name = "test"
+        period = 12 * 3600  # Specify the historic period in seconds
+
+    class HandleRef:
+        tag = "data"
+
+
+class Response(meta.Response):
+
+    """
+    Maintains a cache for third party data responses
+    """
+
+    request = models.OneToOneField(
+        "Request",
+        on_delete=models.CASCADE,
+        related_name="response",
+    )
+
+    class Meta:
+        db_table = "test_meta_respnse"
+        verbose_name_plural = "Response cache"
+        verbose_name = "Response cache"
+
+    class HandleRef:
+        tag = "response"
+
+    class Config:
+        meta_data_cls = Data
+
+
+class Request(meta.Request):
+    """
+    Concrete implementation of Request model
+    """
+
+    target = models.CharField(max_length=255)
+
+    class Meta:
+        verbose_name = "Request"
+        verbose_name_plural = "Requests"
+        db_table = "test_meta_request"
+
+    class HandleRef:
+        tag = "request"
+
+    class Config:
+        meta_data_cls = Data
+        source_name = "test"
+        target_field = "target"
+        cache_expiry = 86400
+
+    @classmethod
+    def target_to_url(cls, target):
+        """
+        override this to handle converting a target to a requestable url
+        """
+        return target  # Simply return the target as the URL